pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tinyrag"
version = "0.1.0"
description = "Tiny RAG starter kit with FastAPI, MCP, and multiple LLM providers"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "Thoughtworks Labs" }
]
keywords = ["rag", "retrieval-augmented-generation", "mcp", "fastapi", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "openai>=1.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.35.0",
    "ollama>=0.5.1",
    "path>=17.1.0",
    "pydash>=8.0.5",
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "aioboto3>=15.2.0",
    "mcp>=1.18.0",
    "numpy>=2.3.4",
    "slowapi>=0.1.9",
    "microeval>=0.1.0",
    "groq>=0.9.0",
]

[project.urls]
Homepage = "https://github.com/boscoh/tinyrag"
Documentation = "https://github.com/boscoh/tinyrag#readme"
Repository = "https://github.com/boscoh/tinyrag.git"
Issues = "https://github.com/boscoh/tinyrag/issues"

[project.scripts]
tinyrag = "tinyrag.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["tinyrag"]

